ref: Move swagger logic to helpers
