+++ src/app.ts
@@ -9,21 +9,17 @@ import session from 'express-session';
 import passport from 'passport';
 import cors from 'cors';
 import helmet from 'helmet';
-import swaggerUI from 'swagger-ui-express';
 import logger from '@tablerise/dynamic-logger';
-import autoSwagger from '@tablerise/auto-swagger';
-import SwaggerDocumentDnD5E from '../api-docs/swagger-doc-dungeons&dragons5e.json';
-import SwaggerDocumentUser from '../api-docs/swagger-doc-user.json';
-
-import RoutesWrapper from 'src/routes/RoutesWrapper';
 import DungeonsAndDragonsRouteMiddleware from 'src/routes/middlewares/DungeonsAndDragonsRouteMiddleware';
 import UserRouteMiddleware from 'src/routes/middlewares/UserRouteMiddleware';
 import ErrorMiddleware from 'src/middlewares/ErrorMiddleware';
+import swaggerGenerator from './support/helpers/swaggerGenerator';
 
 const COOKIE_AGE = 1000 * 60 * 60 * 120;
 const VALID_ENVS_TO_AUTHENTICATE = ['develop', 'prod'];
 
 const app: Application = express();
+const swaggerDocs = swaggerGenerator(process.env.NODE_ENV as string);
 
 app.use(express.json())
     .use(
@@ -37,36 +33,9 @@ app.use(express.json())
     .use(passport.session())
     .use(cors())
     .use(helmet())
-    .use('/health', (req, res) => res.send('OK!'));
-
-if (process.env.NODE_ENV === 'develop') {
-    autoSwagger(RoutesWrapper.declareRoutes()['dungeons&dragons5e'], { title: 'dungeons&dragons5e' })
-        .then((_result: any) => {
-            logger('info', 'Swagger - dungeons&dragons5e - document generated');
-        })
-        .catch((error: any) => {
-            console.log(error);
-        });
-
-    autoSwagger(RoutesWrapper.declareRoutes().user, { title: 'user' })
-        .then((_result: any) => {
-            logger('info', 'Swagger - user - document generated');
-        })
-        .catch((error: any) => {
-            console.log(error);
-        });
-}
-
-app.use('/api-docs/dnd5e', swaggerUI.serve, (req: Request, res: Response) => {
-    const html = swaggerUI.generateHTML(SwaggerDocumentDnD5E);
-    res.send(html);
-});
-app.use('/api-docs/user', swaggerUI.serve, (req: Request, res: Response) => {
-    const html = swaggerUI.generateHTML(SwaggerDocumentUser);
-    res.send(html);
-});
-
-app.use(UserRouteMiddleware)
+    .use('/health', (req, res) => res.send('OK!'))
+    .use(swaggerDocs)
+    .use(UserRouteMiddleware)
     .use(
         VALID_ENVS_TO_AUTHENTICATE.includes(process.env.NODE_ENV as string)
             ? passport.authenticate('bearer', { session: false })
